Started backend design

+++ database/migrations/2020_02_14_141635_create_sessions_table.php
@@ -15,7 +15,7 @@ class CreateSessionsTable extends Migration
     {
         Schema::create('sessions', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('course_id')->unsigned()->index();
+            $table->bigInteger('course_id')->unsigned()->index();
             $table->date('start');
             $table->date('end');
             $table->integer('places');
