+++ database/migrations/2020_02_14_144931_create_comment_replies_table.php
@@ -15,7 +15,7 @@ class CreateCommentRepliesTable extends Migration
     {
         Schema::create('comment_replies', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('comment_id')->unsigned()->index();
+            $table->bigInteger('comment_id')->unsigned()->index();
             $table->integer('is_active')->default(0);
             $table->string('author');
             $table->string('email');
